Feature #1830 bugfixing
git-svn-id: svn+ssh://svn.amanzi.org/var/lib/svn/trunk/awe@4498 0edadee9-8154-4c59-a661-b5f45207fcbd

+++ org.amanzi.neo.core/src/org/amanzi/neo/core/utils/EditPropertiesPage.java
@@ -111,7 +111,7 @@ public class EditPropertiesPage extends WizardPage {
             } else if (nodeType == NodeTypes.SECTOR) {
                 propertyList.add(new PropertyWrapper("ci", Integer.class, "", false));
                 propertyList.add(new PropertyWrapper("lac", Integer.class, "", false));
-                propertyList.add(new PropertyWrapper("beamwidth", Integer.class, "", false));
+                propertyList.add(new PropertyWrapper("beamwidth", Double.class, "", false));
                 propertyList.add(new PropertyWrapper("azimuth", Integer.class, "", false));
             }
             initialize = true;
